@@ -12,40 +12,47 @@
     "clean": "rimraf buildClient buildServer"
   },
   "dependencies": {
-    "@types/react-dnd": "^2.0.34",
-    "@types/react-dnd-html5-backend": "^2.1.8",
-    "@types/react-redux": "^5.0.14",
     "colors": "^1.1.2",
+    "enzyme": "^3.3.0",
     "express": "^4.16.2",
     "express-no-favicons": "^0.0.1",
     "history": "^4.7.2",
+    "lodash": "^4.17.5",
+    "radium": "^0.21.2",
     "react": "^16.2.0",
     "react-dnd": "^2.5.4",
     "react-dnd-html5-backend": "^2.5.4",
     "react-dom": "^16.2.0",
     "react-redux": "^5.0.6",
-    "react-universal-component": "^2.8.1",
+    "react-universal-component": "^2.8.2",
     "redux-thunk": "^2.2.0",
+    "reselect": "^3.0.1",
     "webpack-dev-middleware": "^2.0.4",
     "webpack-flush-chunks": "^1.2.3"
   },
   "devDependencies": {
-    "@babel/cli": "^7.0.0-beta.38",
-    "@babel/core": "^7.0.0-beta.38",
-    "@babel/plugin-proposal-class-properties": "^7.0.0-beta.38",
-    "@babel/plugin-syntax-dynamic-import": "^7.0.0-beta.38",
-    "@babel/preset-env": "^7.0.0-beta.38",
-    "@babel/preset-react": "^7.0.0-beta.38",
+    "@babel/cli": "^7.0.0-beta.39",
+    "@babel/core": "^7.0.0-beta.39",
+    "@babel/plugin-proposal-class-properties": "^7.0.0-beta.39",
+    "@babel/plugin-syntax-dynamic-import": "^7.0.0-beta.39",
+    "@babel/preset-env": "^7.0.0-beta.39",
+    "@babel/preset-react": "^7.0.0-beta.39",
     "@babel/preset-typescript": "^7.0.0-beta.39",
+    "@types/enzyme": "^3.1.8",
+    "@types/lodash": "^4.14.100",
+    "@types/radium": "^0.18.23",
     "@types/react": "^16.0.36",
+    "@types/react-dnd": "^2.0.34",
+    "@types/react-dnd-html5-backend": "^2.1.8",
     "@types/react-dom": "^16.0.3",
     "@types/react-hot-loader": "^3.0.5",
+    "@types/react-redux": "^5.0.14",
     "babel-loader": "v8.0.0-beta.0",
     "babel-plugin-universal-import": "^1.4.0",
     "babel-watch": "^2.0.7",
     "cross-env": "^5.1.3",
     "css-loader": "^0.28.9",
-    "eslint": "^4.16.0",
+    "eslint": "^4.17.0",
     "eslint-plugin-import": "^2.8.0",
     "extract-css-chunks-webpack-plugin": "^2.0.18",
     "prettier": "^1.10.2",
